Add QueryTreeBuilder tests to separate class
